Moved the styles into a styles folder under components.
