+++ app/(tabs)/report.tsx
@@ -1,7 +1,9 @@
 import React, { useState } from 'react';
-import { StyleSheet, View, Text, TouchableOpacity, ScrollView, TextInput, Alert, Switch } from 'react-native';
+import { View, Text, TouchableOpacity, ScrollView, TextInput, Alert, Switch } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { useColorScheme } from '@/components/useColorScheme';
+
+import styles from '../../components/styles/reportStyles';
 
 export default function ReportScreen() {
     const [isRecording, setIsRecording] = useState(false);
@@ -264,160 +266,3 @@ export default function ReportScreen() {
     );
 }
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        padding: 20,
-    },
-    section: {
-        marginBottom: 30,
-    },
-    sectionTitle: {
-        fontSize: 20,
-        fontWeight: 'bold',
-        marginBottom: 15,
-    },
-    voiceButton: {
-        alignItems: 'center',
-        justifyContent: 'center',
-        padding: 30,
-        borderRadius: 20,
-        elevation: 4,
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 2 },
-        shadowOpacity: 0.25,
-        shadowRadius: 4,
-    },
-    voiceButtonText: {
-        color: 'white',
-        fontSize: 18,
-        fontWeight: 'bold',
-        marginTop: 15,
-        textAlign: 'center',
-    },
-    voiceButtonSubtext: {
-        color: 'white',
-        fontSize: 14,
-        marginTop: 5,
-        textAlign: 'center',
-        opacity: 0.9,
-    },
-    formLabel: {
-        fontSize: 16,
-        fontWeight: '600',
-        marginBottom: 8,
-        marginTop: 15,
-    },
-    categoryGrid: {
-        flexDirection: 'row',
-        flexWrap: 'wrap',
-        justifyContent: 'space-between',
-        marginBottom: 15,
-    },
-    categoryButton: {
-        width: '48%',
-        padding: 15,
-        borderRadius: 12,
-        alignItems: 'center',
-        marginBottom: 10,
-    },
-    categoryText: {
-        fontSize: 14,
-        fontWeight: '600',
-        marginTop: 8,
-        textAlign: 'center',
-    },
-    textInput: {
-        padding: 15,
-        borderRadius: 8,
-        fontSize: 16,
-        marginBottom: 15,
-    },
-    textArea: {
-        padding: 15,
-        borderRadius: 8,
-        fontSize: 16,
-        marginBottom: 15,
-        height: 100,
-    },
-    anonymousToggle: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        marginBottom: 20,
-    },
-    submitButton: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'center',
-        padding: 18,
-        borderRadius: 12,
-        gap: 10,
-    },
-    submitButtonText: {
-        color: 'white',
-        fontSize: 18,
-        fontWeight: 'bold',
-    },
-    reportsList: {
-        gap: 15,
-    },
-    reportItem: {
-        padding: 15,
-        borderRadius: 12,
-    },
-    reportHeader: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        marginBottom: 10,
-    },
-    reportTitle: {
-        fontSize: 16,
-        fontWeight: '600',
-        flex: 1,
-        marginRight: 10,
-    },
-    statusBadge: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        paddingHorizontal: 10,
-        paddingVertical: 5,
-        borderRadius: 15,
-        gap: 5,
-    },
-    statusText: {
-        color: 'white',
-        fontSize: 12,
-        fontWeight: '600',
-        textTransform: 'capitalize',
-    },
-    reportDetails: {
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-    },
-    reportCategory: {
-        fontSize: 14,
-    },
-    reportDate: {
-        fontSize: 14,
-    },
-    quickActionsGrid: {
-        flexDirection: 'row',
-        flexWrap: 'wrap',
-        justifyContent: 'space-between',
-    },
-    quickActionButton: {
-        width: '48%',
-        padding: 20,
-        borderRadius: 12,
-        alignItems: 'center',
-        marginBottom: 15,
-    },
-    quickActionText: {
-        fontSize: 14,
-        fontWeight: '600',
-        marginTop: 8,
-        textAlign: 'center',
-    },
-});
